Prof Code Fix

+++ dungeon.cpp
@@ -57,7 +57,7 @@ static void dijkstra_corridor(dungeon *d, pair_t from, pair_t to)
     }
     initialized = 1;
   }
-  
+
   for (y = 0; y < DUNGEON_Y; y++) {
     for (x = 0; x < DUNGEON_X; x++) {
       path[y][x].cost = INT_MAX;
@@ -156,7 +156,7 @@ static void dijkstra_corridor_inv(dungeon *d, pair_t from, pair_t to)
     }
     initialized = 1;
   }
-  
+
   for (y = 0; y < DUNGEON_Y; y++) {
     for (x = 0; x < DUNGEON_X; x++) {
       path[y][x].cost = INT_MAX;
@@ -564,7 +564,7 @@ static void place_stairs(dungeon *d)
            (p[dim_x] = rand_range(1, DUNGEON_X - 2)) &&
            ((mappair(p) < ter_floor)                 ||
             (mappair(p) > ter_stairs)))
-      
+
       ;
     mappair(p) = ter_stairs_up;
   } while (rand_under(2, 4));
@@ -789,7 +789,7 @@ int read_rooms(dungeon *d, FILE *f)
 
       exit(-1);
     }
-        
+
 
     /* After reading each room, we need to reconstruct them in the dungeon. */
     for (y = d->rooms[i].position[dim_y];
@@ -876,7 +876,7 @@ int read_dungeon(dungeon *d, char *file)
     exit(-1);
   }
   fread(&be32, sizeof (be32), 1, f);
-  if (buf.st_size != be32toh(be32)) {
+  if (buf.st_size != (int) be32toh(be32)) {
     fprintf(stderr, "File size mismatch.\n");
     exit(-1);
   }
@@ -885,7 +885,7 @@ int read_dungeon(dungeon *d, char *file)
   d->PC->position[dim_x] = i;
   fread(&i, 1, 1, f);
   d->PC->position[dim_y] = i;
-  
+
   read_dungeon_map(d, f);
   d->num_rooms = calculate_num_rooms(buf.st_size);
   d->rooms = (room_t *) malloc(sizeof (*d->rooms) * d->num_rooms);
@@ -985,10 +985,10 @@ void render_hardness_map(dungeon *d)
 {
   /* The hardness map includes coordinates, since it's larger *
    * size makes it more difficult to index a position by eye. */
-  
+
   pair_t p;
   int i;
-  
+
   putchar('\n');
   printf("   ");
   for (i = 0; i < DUNGEON_X; i++) {
